requirements.txt
streamlit
sounddevice
scipy
openai-whisper
faster-whisper
langchain-groq
python-dotenv
langchain
streamlit-audiorecorder
uvicorn
streamlit-webrtc
fastapi
websockets
streamlit-mic-recorder